requirements.txt
numpy==1.22.3
pandas==1.3.4
pycryptodome==3.14.1
PyQt5==5.15.6
tqdm==4.62.3
